[#510] Implement API endpoints for domains
- Adds routes for domain operations
- Creates separate modules for account and domain logic
- Implements new API classes for handling these operations
- Moves existing account-related logic to the new structure
- Enhances the base API class with common action processing methods

These changes improve code organization and set the foundation for more
robust account and domain management features.
