+++ lib/onetime/app/api/routes
@@ -19,9 +19,12 @@ POST   /v1/metadata/:key                        Onetime::App::API#show_metadata
 
 POST   /v1/private/:key/burn                    Onetime::App::API#burn_secret
 
-POST   /v1/account/destroy                      Onetime::App::API#destroy_account
-POST   /v1/account/change-password              Onetime::App::API#change_account_password
-POST   /v1/account/apikey                       Onetime::App::API#generate_apikey
+POST   /v1/account/destroy                      Onetime::App::API::Account#destroy_account
+POST   /v1/account/change-password              Onetime::App::API::Account#change_account_password
+POST   /v1/account/apikey                       Onetime::App::API::Account#generate_apikey
+POST   /v1/account/domains/add                  Onetime::App::API::Account#add_domain
+DELETE /v1/account/domains/:domain               Onetime::App::API::Account#remove_domain
+GET    /v1/account/domains                      Onetime::App::API::Account#list_domains
 
 ## API v2
 GET    /v2/status                               Onetime::App::APIV2#status
